requirements.txt
# RAG Pipeline Dependencies
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-mongodb>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
pymongo>=4.6.0

# Local tracing with LangFuse v3
langfuse>=3.0.0
